We got a dumb lil animation!!!
- that little glitch on the westward walk cycle happens in bevy as well as godot, now I want to go look and
see if gms was doing it too and I just didn't notice. Something to do with the way the texture gets shipped
to the gpu? Need to re-pack it or something?

- looks like that sprite spawns in the exact middle of the screen if you don't mess with the transform. ok!

- Want some automation to consume the outputs of tutorial boi's asset pipeline, he optimized it for GMS but the
result looks super tractable for really whatever. This is a thing that feels easier to grapple with from a real
language with a real ecosystem than it did in godot, where it felt like I was gonna have to just twiddle widgets
for an eternity. Anyway: need to get the dimensions and divide X by the number in the filename to get what
bevy's TextureAtlas::from_grid wants.

- Also next: split the cardinal dirs and react to input!