OK wow, I got that working! It stutters nastily, so probably need to finally fix my timestep next. Also:
- I initially was grabbing the translations off the transforms and mutating those, but didn't pay attention to my
type hints: Vec3 implements Copy, and so assigning them to a temporary var won't let you mutate the original
(you'd have to assign it back to the struct field at the end.) So obviously that didn't work. Owned.

- with a queryset, you need to finish using each sub-query before using the next! Well -- back to Vec3 being
Copy, I can just grab a copy of the info I need and let non-lexical lifetimes end that borrow before moving on to
the camera query.

- I would rather use get_single_mut than iterating, but there was some kind of issue where the borrow
ended too early. Now that I think about it, I think I know how to deal with that (capture the reference
returned by `q1()`), but anyway, this is working.