Switch window to FIFO present mode, lock delta timestep to 1/60 secs
I started getting weird stutter and jank AGAIN, and, let's see:

1. The mailbox present mode actually runs WAY ahead of the
vblank and it turns out it was spitting delta times of like, idk,
0.0025 seconds? instead of 0.0166? Like 7x as fast! So it should have been
smooth as butter, but it was janking out, probably because of delta
being fundamentally inaccurate!

2. I've started wondering if my particular system showcases this more, because
while it's got a reasonably powerful (tho modest) graphics card, the mobo is
stuck on pcie 2.0 (or worse?). Perhaps the difficulty in pushing bits over the
wire amplifies the inaccuracy of frame time measurements!

3. I decided fuck it, let's see
what happens if we don't trust delta time AT ALL.
Just switch back to FIFO present-mode (locked to vblank), and pretend like
delta is always exactly 1/60th (this is what gamemaker does, I think?). If we
miss our frame we get slowdown, but the timings on mailbox mode seem to
imply that making our frame shouldn't be a challenge.

I'm still using the .delta_seconds() interface for everything here,
so it should be easy to rip this back out later. I've got some other ideas
too, maybe. (Time to try for implementing fixed timestep on top of delta?)