Okay hey, this works! For what it is! (showing first frame)
The atlas replacement system fires when expected and only when expected
(though it's doing the comparison constantly, which seems a waste, but,
exploiting change tracking and/or asset events comes later!!). Right now
it's displaying the first frame of the whole atlas, because the  default
sprite sheet bundle sets index to 0.

Next: idk, respect the frame index or something.