Implement TryFrom<&str> for Dir
(this is only half a stride of work as I actually did it, but it stands alone.)

Wow!! Ok that was a lot of hoops to jump through to make the error
result questionmarkable (in a function that returns an anyhow::Error) (I
originally tried just setting type Error = String and it hated that) but!!
this lets me do `let thing: Dir = "N".try_into()?;`, so now I can refactor
char_animation to use an enum for directions instead of strings. hell yeah.