Never mind! Instead of storing size, let's pre-compute the anchor point
Hey, baby's first affine transform! What I learned: affine transforms
are a superset of linear transforms, where you can add a _translation_
in addition to the scale/shear/rotate. In bevy/glam, we can build one
by combining a linear transform (n x n matrix) and a translation vector.

Still trying to get my head around why... OH NO WAIT, I GOT IT. lemme
draw a picture. We start with pixel coords, so origin at top left.

               x       ,    ┌──,
┌──►          x       x     │ x
│    ┌────┐  ┌────┐  ┌────┐ ▼┌────┐
│    │x   │  │    │  │    │  │    │
▼    │ x  │  │    │  │    │  │    │
     │    │  │    │  └────┘  └────┘
     │    │  │    │
     │    │  │    │
     └────┘  └────┘

(https://asciiflow.com/#/)

Ok, so,

- Reflect Y.
- Normalize the scale,
- Translate by adding on that (-.5, .5) vector.

At no point do you change the coordinate system in here, yet, and we're
still dealing with "y goes down". So at the end, you've got a vector
with a tiny bit of positive Y (below the top of the box), and a tiny bit
of negative X. That goes nowhere near the point that the original vector
would have landed at.

But!! If you throw all that away and start with your origin at the
center of the rectangle, and with coordinates set to plus/minus a half
at the edges of the box, _and then apply that nonsense vector we just
built!_ Then that resulting point ends up at the same position relative
to the box edges as the original point!

Ugh, linear algebra is hard.