Add TempCharAnimationState::ChangeAnimation
Zero out the frame and timer to keep state clean-ish. Not sure what to
do about variant count yet.

Originally I thought the last commit might leave a gap on sprite change, but
there's a case to catch that when the frame_timer is None... so, just
got to make sure that's true whenever the animation has changed.

Oh, and system order is starting to matter here. :|