I solved the mystery of the inside-out rectangle!!!
I initially thought the vec flip was the problem and that I needed to
take the height into account, but actually that was fine because I was
already moving the rectangle to be relative to the right origin point
before doing that. No, the problem was that I was accurately translating
the upper-left and bottom-right corners into bottom-up coordinates, but
those ARE NOT THE MIN/MAX CORNERS of that rectangle in the new regime!
You want bottom-left and upper-right.

With that fixed, the walkbox/wall collisions work exactly like they used to!