Make charanm_test_directions_system less dumb (but still p dumb)
This uses "real" direction crunching so it feels way nicer! It's still
the wrong place for this and still ignores gamepad.

Fun bug it smoked out: change_variant constantly resetting frame to 0
while moving, so we moonwalk while standing still, but stand still while
moving. The moonwalk was already silly, but that really took it up a
notch!