OK, I think that sorts that panic out --
I got a crash and a panic backtrace like this:

```
   0: bevy_ecs::schedule::executor_parallel::system
           with name="bevy_sprite::render::extract_sprites"
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_ecs-0.8.0/src/schedule/executor_parallel.rs:182
   1: bevy_ecs::schedule::executor_parallel::prepare_systems
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_ecs-0.8.0/src/schedule/executor_parallel.rs:170
   2: bevy_render::stage
           with name="extract"
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_render-0.8.0/src/lib.rs:235
   3: bevy_render::renderer subapp
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_render-0.8.0/src/lib.rs:211
   4: bevy_app::app::frame
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_app-0.8.0/src/app.rs:118
   5: bevy_app::app::bevy_app
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_app-0.8.0/src/app.rs:131
thread 'Compute Task Pool (3)' panicked at 'index out of bounds: the len is 4 but the index is 4', /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_sprite-0.8.0/src/render/mod.rs:273:29
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
   0: bevy_render::stage
           with name="extract"
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_render-0.8.0/src/lib.rs:235
   1: bevy_render::renderer subapp
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_render-0.8.0/src/lib.rs:211
   2: bevy_app::app::frame
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_app-0.8.0/src/app.rs:118
   3: bevy_app::app::bevy_app
             at /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/bevy_app-0.8.0/src/app.rs:131
thread 'main' panicked at 'task has failed', /Users/nick/.cargo/registry/src/github.com-1ecc6299db9ec823/async-task-4.2.0/src/task.rs:425:43
```

That doesn't trace the problem all the way back to my app code, but it
does at least tell me there's a TextureAtlasSprite that's trying to
index past the length of a TextureAtlas.

The length of the idle atlas is 4 (one frame, four directions), and the
length of the run atlas is much longer. What that sounds like to me is,
we ended a frame in an inconsistent state -- the atlas was still set to
"idle", but the sprite had an index value intended for "run".

- `charanm_atlas_reassign_system` is what changes the atlas (to match the
  current value from the `CharAnimationState`).
- `charanm_animate_system` is what changes the sprite (to match the current
  value from the `CharAnimationState`).
- Main game logic systems are allowed to mutate the `CharAnimationState`.
- The current scheduling constraints allowed the texture_reassign and
  animate systems to be separated by arbitrary app systems in the
  default stage (core update). This order is not guaranteed to be
  deterministic (even across app runs, I think? Certainly not across
  compiles).

So, I think we found our killer.

The correct move in current staged ecs lifestyle is to make your own
dedicate stage for a plugin like this, but it's in the same repo as my
app code and I'm the only one using it, and that lifestyle isn't long
for this world, so it doesn't seem worth that effort. I'll just use
PostUpdate and say you can't touch the animation state after Update. (Or
if it comes down to it, I'll resort to Last I guess.)