Ok!! Baby's first state transition! It sucks but it works!
Right, dealing with dynamic types in anything less than a whole-hog
Reflect-driven way turned out to be a nonstarter. And doing this
generically is ass because of the need to add generic versions of each
system to the app. So!! For the time being, I'll just store the "done"
condition on the state itself, and I'll rep it as an enum so I can cram
the next state directly in there. The clone derive turns out to be
crucial for getting it back _out_ again... oh, although, actually I
think mem::swap is good enough for this. But probably not much faster.
So I'll leave that be for now.

Anyway! This transitions out of roll into free! Now to go the other way
around... and then I need to fix the roll animation speed.