Migrate all obvious code to handle Bevy 0.9 changes
It compiles and runs! The only fly in the ointment is that the colliders are
offset somehow. I have a distinct suspicion that the issue is in bevy_ecs_ldtk
rather than in my character walkboxes, but let's see if we can spin up some
debug code to efficiently prove it!