Fix "half-price kitty pryde" collision check bug
h/t to Brenna for the joke, ANYWAY ldtk IntGridCell entities are children of a
layer entity that is now, as of bevy_ecs_ldtk 0.5, being translated half a tile
up and right, putting their local Transforms in a totally different reference
frame from the player ((8,8) vs. (0,0)). That's why the debug mesh (a child of
the entity) was rendering in the right spot but the collision check was looking
somewhere entirely different.

For immobile solids, the solution is obvious, but it's gotten me worried about
mobile ones! But we'll cross that bridge when IT CROSSES US.