Refactor movement systems to produce/consume velocity
This runs! It's in a bit of an in-between state, though. The bonk and roll
states are still using a total distance budget to handle their transitions, so
they're still reading Time and getting a redundant total movement vector.

I think this change exacerbates any inconsistency that comes out of a non-fixed
timestep, but for the time being I think I'm going to just have to put up with
that. I'll keep it in the back of my mind, but I'm not getting networking for...
a long time, at this point. So oh well!!

Oh btw, I don't think I mentioned this, but Isaac pinpointed what was probably
the source of the "eternal bonk" bug: it was subtracting the length of the
_clamped_ movement vector from the total distance budget, rather than the length
of the raw one. So I've switched back to chipping away at an f32, and am
subtracting the raw movement instead. (Which, incidentally, now the raw movement
is the ONLY movement, cf. what I was saying above about variable timestep
inaccuracies.)