Put char_animation systems in a set, position them in Update between plan and move systems
OKay, I drew a messy diagram of this in my notebook which I should probably
record here in this repo somehow, but basically having the animation steps all
happen in PostUpdate is WRONG. It causes the action of a single game tick to
smear out across multiple frames.

Main upshot is that:

- `free_plan_move` produces DIRECTION.
- All the move planners produce SPRITE CHANGE under some conditions.
- Animations systems consume DIRECTION and SPRITE CHANGE, and produce POSITIONAL DATA
  retrieved from the animation frame assets.
- Move system consumes POSITIONAL DATA for collision.

I think a sort-of okay API for this is to let the animation plugin publish a
SystemLabel (currently called `SpriteChangers`) and expect the app to apply that
label to any system that changes sprites _for the current frame._ Might
reconsider later! But for now, also using a similar trick for move planner
systems that need to go before the planned move system.