Switch frame timer to TimerMode::CountUp, and recycle excess frametime
Figuring out how to sell that while loop to the borrow checker was a bit of a
puzzle, but I'm satisfied with the result.

Anyway!! This uses an unmerged timer API of my own devising that acts like my
kitchen timer does -- it counts up like a stopwatch after it finishes its
countdown. (Alice said this seemed "well-motivated" so I'll PR once I've shaken
out any kinks.)

I'm using this to follow up on a logic flaw that Isaac pointed out: if there's
any delta time _in excess_ of what it took to flip the frame timer over, you
want to roll it directly into the next frame's countdown, because otherwise your
animations will streeeeetch if you're in an imperfect resonance with the vsync
rate.

I might be imagining things, but between this and the one-frame coherency
ordering fixes I just did... this feels smoother?! Man. Gamedev's hard.