Fix one-frame lag on walkbox updates (WHOOPS)
This didn't have a chance to become visible because my character walkboxes stay
a constant size. Also, as soon as I start fucking with different layouts, I'm
going to have a problem with getting stuck inside walls, but NEVER MIND that for
now. Anyway: Commands are deferred until the end-of-stage sync point, and the
animation systems live in the update stage now in-between some other gameplay
systems.

As part of this I had to sort out what to do about missing walkboxes, and I
decided to just shrink them to 0px, without really thinking all the way through
what that'll do. It's probably fine!! (not.) But the alternative was to switch
to `Walkbox(Option<Rect>)` and I kind of hated that.