Respect FrameTimeOverride in subsequent frames, not just the first one
W h o o p s! Longroll was always a bug! This fixes it -- we get the actual
specified roll distance, and with ALL the frames of animation instead of
stopping a little over halfway through.

For convenience, this is implemented as a method on variant, because to do the
full calculation you need access to both the frame itself and the total number
of frames.