Make MotionKind::Continuous the new default since I'm currently working on it
By the way, I think I've figured out part of the problem with my zips: it's
because I'm trying to do x and y faceplants at the same time per box, with only
one iteration through the boxes, but I'm doing the complete X and then the
complete Y truncations in the per-box work.

I didn't word that very usefully but: by doing a complete X exclusion, you're
now guaranteed to be outside of this box! So we're never going to run the Y part
for this box! But now you're planning to go inside _a different box,_ and if
it's later in the iteration it'll push you out of its X boundaries as well (but
not fuss w/ the Y because you're no longer it's problem.)

I think one answer might be to try and clamp the proposed vec2 proportionally,
so you're sort of doing x and y simultaneously. Lemme look at the vector methods
and see what occurs to me.