Bah ok I can't construct a Cursor at all outside default().
(Though the public fields are mutable if need be.)

By the way! With this, everything actually seems to be working EXCEPT the
space-flattener system, which is fucked because we lost translation_mut() on
GlobalTransform. But, I think we can work around that, and outside the extract
schedule, too — after all, I just went to the trouble of adding that whole
PhysTransform system! But, I'll have to expand it and I think I'll have to make
almost everything adopt it, including the shadows themselves.

But yeah anyway my theory is that since PhysTransform gives us an effectively
global reference frame anyway, we should be able to use it in place of
GlobalTransform for this.