Fix TopDownMatter Z-sync!
HEYYYYY, that's the last thing that's busted from the upgrade! Incredible!

I was so sure I was going to have to do something outrageously baroque to get
around the lack of `get_translation_mut()`, but holy shit, the stupid path of
just constructing a new GlobalTransform from scratch and swapping it into place
worked great.