Add X-flipping for one-sided sprites... sort of!
- At the time of loading, the animation asset as a whole needs to remember how
many variants it's got and how they're configured (I'm using a new
Directionality enum for this).
- Then, the animation state needs to know whether its sprite is currently
flipped, and we need to update this along with the variant ID as part of setting
the direction.
- Then, we need to propagate that to the TextureAtlasSprite when we animate.

This works! However, it looks like we're gonna need to also mirror the anchor
point and the collider boxes. Fiddly-ass shit!!