TEMP: turn player into slime, and break animation slightly?
I want to test the mirrored behavior as I try and sort out the origin/collider
mirroring, but also, check this out — in this condition, if the player is facing
any direction but east and starts a roll, their animation is fucked up! Way too
few frames! They seem to go the right distance.

What could be causing this???? Where do I look first?