Fix anchors and colliders on x-mirrored sprites!!
- Flip is determined and recorded in set_directions_system
- animate_system sets anchor point
- update_colliders_system sets walk/hitboxen
- both of those are downstream of set_directions, in a given frame.

Refresher on sprite anchor format: the texture's rectangle gets squished and
normalized such that center is 0,0 and top left is -.5,.5. So, we can mirror
around the center just by flipping the X sign!

And, the flip_rect functions mirror the rectangle around wherever the corners
are positioned relative to. so if it were something like:

____
|   |
-----  x

You'd get,

    ____
   |   |
x  -----

which is exactly right.