ldtk level: rearrange water/grass auto tile rules... maybe
This changes the borders of the colliders for water tiles to be... on the marked
water tile. I don't know if I like it tbh! These should be HALF or QUARTER
colliders. But, that takes some more work and I need to set up like tile enum
tags or something. Distraction!