Fix z-height render system. BUT.....!!!
They changed the layout of this, but actually I think the new layout lets me get
around the excessive iteration! So come back to this.

At this point, all red errors are banished and the game compiles and runs as
expected! Something around the speeds seems slightly wonky, and I'm gonna want
to correlate with main to see if I can spot what's up. (Kitty seems slow, and
Sloom seems to sometimes accelerate weirdly? Or get stuck?)