Refactor move system
Before, we were iter_mut()ing over the whole query and maintaining a separate
hashmap for cross-referencing with other entities we might collide with.

Now, we're extracting the entity IDs, sorting them by PushPriority, iterating
that slim list, and grabing the current mover out of the query one-by-one.

This should hopefully be more or less a wash, and tracy shows the system as
still averaging single-digit microseconds.