New omnibus collider debug updater w/ hurtbox support
I did a first attempt of this where a constructor function would generate
generic closure systems for each box type, but wow, it was heinous! Doing a
single system with a ParamSet and a helper function is MUCH cleaner, and
practically speaking there's no downside because those separate systems would
have had to run in serial anyway due to the overlapping mutable access.