char_animation updates: wow, texture atlas sure changed a lot! (more)
To show an animated texture atlas sprite, you need four buckets of stuff:

- A list of rectangles for frames
- The current frame index
- The actual images, packed into an omni-texture
- Some misc crap

Crucially, non-animated sprites also need a `Handle<Image>` and some misc crap.
So in 0.13 they unified the storage locations of those into a bare
handle-as-component and a `Sprite`, and moved the atlas stuff to accommodate.
Main thing to remember is that none of the TextureAtlas stuff has a pointer to
an omni-texture handle anymore; it's stored separately and has to be queried in
addition.

```
// Bundle has a special type of sprite and a Handle<TextureAtlas>:
pub struct SpriteSheetBundle {
    pub sprite: TextureAtlasSprite,
    pub texture_atlas: Handle<TextureAtlas>,
    // ...
}

// Same as Sprite component, except it has an `index` instead of a `rect`:
pub struct TextureAtlasSprite {
    pub index: usize,
    // ...
}

// TextureAtlas is an Asset, but also Handle<TextureAtlas> implements Component.
// There's a Handle<Image> embedded in an atlas asset!
pub struct TextureAtlas {
    pub texture: Handle<Image>,
    pub size: Vec2,
    pub textures: Vec<Rect>,
    /* private fields */
}
```

```
// Now it's much closer to the non-animated SpriteBundle -- the only unique item
// is the TextureAtlas component! TextureAtlasSprite is gone-zo.
pub struct SpriteSheetBundle {
    pub sprite: Sprite,
    pub texture: Handle<Image>,
    pub atlas: TextureAtlas,
    // ...
}

// TextureAtlas is a Component that stores the mutable state for animation (it's
// just the index) and a layout handle. TextureAtlasLayout is an Asset with the
// static rectangle deets. NONE of these things gets to hold a Handle for the
// actual texture image; that's stored alongside it as a top-level component.

// COMPONENT
pub struct TextureAtlas {
    pub layout: Handle<TextureAtlasLayout>,
    pub index: usize,
}

// ASSET
pub struct TextureAtlasLayout {
    pub size: Vec2,
    pub textures: Vec<Rect>,
    /* private fields */
}
```

In short: you give the bundle what it needs, but for an entity that might have
multiple animations over its lifetime, you're gonna need to dink with the
`TextureAtlas` component, the `Handle<Image>` component, and
`TextureAtlasLayout` assets.